+++ build.rs
@@ -33,7 +33,7 @@ fn main() {
             .user_type_name("RaaTtGrammar")
             .user_trait_module_name("raa_tt_grammar")
             .trim_parse_tree()
-            .disable_recovery()
+            // .disable_recovery()
             .generate_parser()
         {
             ParolErrorReporter::report_error(&err, "raa_tt.par").unwrap_or_default();