+++ src/raa_tt_parser.rs
@@ -412,7 +412,6 @@ where
         NON_TERMINALS,
     );
     llk_parser.trim_parse_tree();
-    llk_parser.disable_recovery();
     let scanner = RaaTtGrammarScanner::new();
     // Initialize wrapper
     let mut user_actions = RaaTtGrammarAuto::new(user_actions);